Cleanup docker containers on delete
Use remove_container rather than kill to remove the container. Kill is a force stop of the container, but it leaves the container around. remove_container with force=True is the way to stop and actually remove the container.

Leaving containers around in a stopped state is a general issue, but specifically if named containers are being used (e.g. to be able to link them together) then after a delete, the same containers cannot be recreated because they already exist by the same name.